+++ src/shapes.js
@@ -26,3 +26,12 @@ export const departmentsShape = PropTypes.arrayOf(PropTypes.shape({
   code: PropTypes.string.isRequired,
   name: PropTypes.string.isRequired,
 }));
+
+export const affiliationsShape = PropTypes.arrayOf(PropTypes.shape({
+  id: PropTypes.string,
+  isPrimary: PropTypes.bool,
+  tenantId: PropTypes.string,
+  tenantName: PropTypes.string,
+  userId: PropTypes.string,
+  username: PropTypes.string,
+}));